Criando gravação do procedimento.

+++ src/main/kotlin/br/unipar/plano/domain/procedimento/usecases/impl/CancelaProcedimentoUseCaseImpl.kt
@@ -1,19 +1,20 @@
 package br.unipar.plano.domain.procedimento.usecases.impl
 
-import br.unipar.plano.domain.centrais.model.CentralRepository
-import br.unipar.plano.domain.centrais.model.IdCentral
-import br.unipar.plano.domain.centrais.usecases.DeletaCentralUseCase
 import br.unipar.plano.domain.procedimento.model.IdProcedimento
-import br.unipar.plano.domain.procedimento.model.Procedimento
 import br.unipar.plano.domain.procedimento.model.ProcedimentoRepository
-import br.unipar.plano.domain.procedimento.usecases.DeletaProcedimentoUseCase
+import br.unipar.plano.domain.procedimento.model.StatusProcedimento
+import br.unipar.plano.domain.procedimento.usecases.CancelaProcedimentoUseCase
 import org.springframework.stereotype.Service
 
 @Service
-class DeletaProcedimentoUseCaseImpl(private val procedimentoRepository: ProcedimentoRepository) : DeletaProcedimentoUseCase {
+class CancelaProcedimentoUseCaseImpl(private val procedimentoRepository: ProcedimentoRepository) : CancelaProcedimentoUseCase {
 
     override fun executa(idProcedimento: IdProcedimento) {
         val procedimento = procedimentoRepository.findById(idProcedimento).orElseThrow { ProcedimentoNotFoundException(idProcedimento) }
+
+        if (procedimento.status.equals(StatusProcedimento.CANCELADO))
+            throw IllegalStateException("O procedimento com id ${procedimento.id} já foi cancelado!")
+
         return procedimentoRepository.delete(procedimento)
     }
 }
